feat: important update to 0.5.0
- Remove all content about clash and ssh port forwarding. Internet proxy will be directly set through the ip address.
- The web service will be provided by another package slash.
